Show subforums under forums on main board
Subforums will be displayed underneath their parents with an indent.
The code for this was posted on gist:
    https://gist.github.com/phase/ba2df6805ab90dac746d
That only contains the subforum.tpl, but it's the only piece of code
that matters. Some CSS changes have been made since the posting of it.

The diff for layout.php is so large because of all the leading
whitespace that was left in the file.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>